Ignore any errors when interpolating
When there's an error validating the response then Pink Pony will
return an errors object that contains more detailer errors.

Example:
  {
    errors: [
      {details: "This key is missing"}
    ]
  }

The interpolation expects a string and this clearly isn't a string.

We wanted to get the fix out and we'll do an integration test that will
get Spring to insert the wrong value for us as the type system doesn't
let us do this in unit.

[#116592631]